// apps/backend/models/assignment.model.ts
import mongoose from "mongoose";

const assignmentSchema = new mongoose.Schema(
  {
    teacherId: {
      type: mongoose.Schema.Types.ObjectId,
      ref: "User",
      required: true,
    },

    title: {
      type: String,
      default: "Untitled Assignment",
    },

    subject: { type: String, required: true },
    className: { type: String, required: true },
    timeAllowed: { type: Number, required: true },

    dueDate: {
      type: Date,
      required: true,
    },

    materialUrl: {
      type: String,
      default: null,
    },

    pdfUrl: {
      type: String,
      default: null,
    },

    jobError: {
      type: String,
      default: null,
    },

    bullJobId: {
      type: String,
      default: null,
    },

    questionConfig: [
      {
        questionType: { type: String, required: true },
        count: { type: Number, required: true },
        marks: { type: Number, required: true },
      },
    ],

    additionalInstructions: {
      type: String,
      default: null,
    },

    // BullMQ background job state
    status: {
      type: String,
      enum: ["PENDING", "PROCESSING", "COMPLETED", "FAILED"],
      default: "PENDING",
    },

    pdfStatus: {
      type: String,
      enum: ["NONE", "PENDING", "PROCESSING", "COMPLETED", "FAILED"],
      default: "NONE",
    },

    // The final JSON output generated by the LLM
    generatedPaper: {
      totalMarks: {
        type: Number,
        default: 0,
      },
      sections: [
        {
          title: { type: String }, // e.g., "Section A"
          instructions: { type: String }, // e.g., "Attempt all questions"
          questions: [
            {
              text: { type: String },
              difficulty: {
                type: String,
                enum: ["Easy", "Moderate", "Challenging"],
              },
              marks: { type: Number },
              answer: { type: String },
            },
          ],
        },
      ],
    },
  },
  {
    timestamps: true,
  },
);

const Assignment = mongoose.model("Assignment", assignmentSchema);

export default Assignment;